// app/commons/share_app/components/ReadMail/Attachments/types.ts
export type GroupName = 'application' | 'audio' | 'image' | 'text' | 'video';

export type FileType =
  | 'default'
  | 'msword'
  | 'pdf'
  | 'vcard'
  | 'vnd.ms-excel'
  | 'vnd.ms-powerpoint'
  | 'vnd.oasis.opendocument.text'
  | 'vnd.oasis.opendocument.spreadsheet'
  | 'vnd.oasis.opendocument.presentation'
  | 'vnd.openxmlformats-officedocument.presentationml.presentation'
  | 'vnd.openxmlformats-officedocument.spreadsheetml.sheet'
  | 'vnd.openxmlformats-officedocument.wordprocessingml.document'
  | 'x-7z-compressed'
  | 'x-compressed'
  | 'x-gzip'
  | 'x-zip-compressed'
  | 'zip'
  | 'html'
  | 'epub+zip'
  | 'json'
  | 'octet-stream'
  | 'wav'
  | 'mpeg'
  | 'ogg'
  | 'x-vcard'
  | 'quicktime'
  | 'avi'
  | 'mp4'
  | 'webm'
  | 'x-ms-wmv'
  | 'x-msvideo'
  | 'jped'
  | 'bmp'
  | 'png'
  | 'webp';